API and doc cleanup
