+++ Sources/AutomergeRepo/Documentation.docc/Curation/PeerToPeerProviderConfiguration.md
@@ -4,7 +4,7 @@
 
 ### Creating a peer-to-peer provider configuration
 
-- ``init(passcode:reconnectOnError:autoconnect:recurringNextMessageTimeout:waitForPeerTimeout:)``
+- ``init(passcode:reconnectOnError:autoconnect:logVerbosity:recurringNextMessageTimeout:waitForPeerTimeout:)``
 
 ### Providing a default name for peer-to-peer network advertising
